+++ libc/string.h
@@ -24,15 +24,15 @@ typedef enum {
     DEC
 } num_sys_t;
 
-bool memcmp(const void *s1, const void *s2, size_t len);
-void memset(void *addr, uint8_t val, size_t len);
-void* memcpy(void *dest, const void *src, size_t len);
+bool memcmp(const void *s1, const void *s2, uint64_t len);
+void memset(void *addr, uint8_t val, uint64_t len);
+void* memcpy(void *dest, const void *src, uint64_t len);
 
 char *strcpy(char *__restrict dest, const char *src);
-char *strncpy(char *__restrict dest, const char *src, size_t len);
+char *strncpy(char *__restrict dest, const char *src, uint64_t len);
 int strlen(const char *s);
 int strcmp(const char *a, const char *b); 
-int strncmp(const char *a, const char *b, size_t len);
+int strncmp(const char *a, const char *b, uint64_t len);
 int strcat(char *dest, const char *src);
 char *strchr(const char *s, int c);
 uint64_t strtol(char* s, num_sys_t type);
